// src/common/enums/book-genre.enum.ts
// src/common/enums/book-genre.enum.ts

export enum BookGenre {
  FICTION = "fiction",
  NON_FICTION = "non-fiction",
  SCIENCE = "science",
  HISTORY = "history",
  BIOGRAPHY = "biography",
  ROMANCE = "romance",
  MYSTERY = "mystery",
  THRILLER = "thriller",
  HORROR = "horror",
  FANTASY = "fantasy",
  SCIENCE_FICTION = "science-fiction",
  YOUNG_ADULT = "young-adult",
  CLASSIC = "classic",
  POETRY = "poetry",
  COMICS = "comics",
  SELF_HELP = "self-help",
  BUSINESS = "business",
  TRAVEL = "travel",
  RELIGION = "religion",
  ART = "art",
}